add lab1

+++ lab3.c
@@ -3,38 +3,38 @@
 
 int main()
 {
-    float a = 0;
-    float x;
+    double a = 0;
+    double x;
 
     printf("Enter x step for argument -> ");
-    scanf("%f", &x);
+    scanf("%lf", &x);
     printf("DO WHILE CYCLE\n   x\t\tf(x)\n");
     do
     {
-        printf("%f\t%f\n", a, exp(sin(a)));
+        printf("%lf\t%lf\n", a, exp(sin(a)));
         a += x;
     } 
     while (a <= 0.250001f);
 
     do
     {
-        printf("%f\t%f\n", a, exp(a) - (1 / sqrt(a)));
+        printf("%lf\t%lf\n", a, exp(a) - (1 / sqrt(a)));
         a += x;
     } 
-    while (a <= 0.500001f);
+    while (a <= 0.500001);
 
     printf("Enter x step for argument -> ");
-    scanf("%f", &x);
+    scanf("%lf", &x);
     printf("FOR CYCLE\n   x\t\tf(x)\n");
-    for(int j = 0; j*x <= 0.500001f; j++)
+    for(int j = 0; j*x <= 0.500001; j++)
     {
-        if(j * x <= 0.250001f)
+        if(j * x <= 0.250001)
         {
-            printf("%f\t%f\n", j*x, exp(sin(j*x)));
+            printf("%lf\t%lf\n", j*x, exp(sin(j*x)));
         }
         else
         {
-            printf("%f\t%f\n", j*x, exp(j*x) - (1 / sqrt(j*x)));
+            printf("%lf\t%lf\n", j*x, exp(j*x) - (1 / sqrt(j*x)));
         }
     }
